Add test rig for automating mongo tests
Adds test code for automating the mongodb portion of tests. Set up so
the tests can be run as part of kb-sdk test or standalone (e.g. in
eclipse). Defaults for tests properties are set so they work with kb-sdk
test.

Note that the the container hangs after the BUILD SUCCESSFUL message.

TODO:
Automate the rest of the mongodb related tests
Add an ElasticSearch controller
Automate the ES related tests
Figure out why the java process hangs after a successful test run.