WIP fix script complication fails
The script complication circuit breaker activates *a lot* when cloning a
workspace (cloning code is in another branch ATM). This fixes 3/4
scripts to only need to compile once when first used, rather than
possibly, depending on the parameters, on every invocation.

The 4th script can't easily be fixed the same way since it is
dynamically generated and will need to be recompiled for each different
generation event.

However, the tests now pass.