Refactor child event handling
The changes here, although numerous, are actually pretty simple and just
entail wrapping child StatusEvents in ChildStatusEvent which associates
the parent ID with the child event.

There is currently no reasonable test suite for the
WorkspaceEventHandler or the IndexerWorker and building one would take a
really long time so manually testing for now. The only real logical
changes (other than removing code) are some minor logging changes.

Child events are no longer stored in the database, but these changes
make it pretty easy to do so.