Clean up diverging methods
There are currently two indexObject?() methods in
ElasticIndexingStorage, oncfor a single object and one for an object and
sub objects. Only the latter is actually used in the indexing code; the
former is only used for tests.

The method implementations were starting to diverge as changes were made
to the 'real' method, so the code was changed so the single object
method now delegates to the multiple object method.

This means there is a performance hit since the multiple object method
writes a temp file and the single object method does not, but this only
affects tests.