Add exclude subobjects plumbing
Added a number of hashCode() and equals() without tests in order to be
able to mock SearchMethod's dependencies. In order to use EqualsVerifier
(and I don't proposed to write tests by hand) all the object fields need
to be final, which means writing builders for about 5 objects with lots
of parameters. Since the methods are generated by eclipse, I'm not going
to worry about this for now.