Index sub objects before std objects
If we switch to checking data object guids in the elasticsearch index
rather than access document guids to ensure an object is indexed,
threre's a race condition that the data object guid may exist while the
subobjects are still indexing. This ensures the subobjects go first and
so if the guid exists, indexing is complete.