Treat elastic conflict exceptions as non-fatal
They're not great in that they mean the code isn't handling indexing
correctly in some cases, but shutting the worker down and leaving events
in a stuck state is worse.

This doesn't fix the root cause of the problem but means there's no
known errors that cause a worker shutdown where the worker continuing
does no further harm.