Add error type to errors
Adds an error type to errors that reach the portion of the worker code
where events get updated in the DB. Currently does nothing, but future
PRs will add the error type to the DB on a failed event. This will allow
for filtering on the error type when doing error analysis, rather than
having to filter on arbitrary error strings.