Fix error when no subobjects are found
When a subobject array in a parent object is empty, the ES code will
insert a record with no `key.*` properties with the parent guid in the
index for the type. E.g. if a genome has no features (which really
should never happen, but there are examples in the wild), ES will insert
a record with the parent object's GUID in the genomefeature index, which
is bad, because searches for the parent object will find it and it'll
have the wrong type, which can screw up recursive indexing in some
cases.

The reason ES does this is to support specs with no indexing rules, such
that the general object properties of a type can be indexed even if
we're not interested in the contents. We want the behavior for parent
objects, but not subobjects.